requirements.txt
asgiref==3.5.0
ccm==3.1.5
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
gunicorn==20.1.0
numpy==1.22.3
pandas==1.4.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
whitenoise==6.0.0
